Split test_device.py into multiple modules
Move device.base tests into separate module

Move dev.channels tests into separate module

Move device.exg tests into separate module

Move test_device.py to dev/test_device_fw_version.py
